Set Wallie V 2.0
Android App created for fun and entertainment purspose. It lets user click instant mood images from camera and sets it as their wallpaper. It does not let's user save images as it will turn out into a memory wastage. So the idea is if user wants to change wallpaper daily, he/she does not needs to download the image daily. Just open the app,click the image,apply filter and set it as wallpaper. Also added a login screen for more Interaction of user..!!